requirements.txt
fastapi[all]~=0.94.0
pydantic~=1.10.6
SQLAlchemy~=2.0.0
alembic~=1.10.2
passlib[bcrypt]~=1.7.4
python-jose[cryptography]~=3.3.0
pytest~=7.2.2
captcha~=0.4